fix: derive worker-health alert from the live run so it clears on start
The "No worker processing this run" banner was computed once in the page load
and never recomputed, so it lingered after a worker actually started and began
heartbeating. Move the check into a client `$derived` over the live run (it also
runs during SSR for the initial paint, and recomputes each SSE tick), and stream
the run's `requestedAt` so the queued-grace check works client-side.
