+++ src/routes/dashboard/stream/+server.ts
@@ -31,6 +31,7 @@ function serialize(r: SyncRun) {
 		bytesStored: r.bytesStored,
 		bytesEstimated: r.bytesEstimated,
 		currentUrl: r.currentUrl,
+		requestedAt: r.requestedAt?.getTime() ?? null,
 		heartbeatAt: r.heartbeatAt?.getTime() ?? null,
 		startedAt: r.startedAt?.getTime() ?? null
 	};
